refactor: move OscSender implementation to a separate module
